Organize the interfaces
