+++ Guppi.Console/Skills/UtilitiesSkill.cs
@@ -4,7 +4,7 @@ using System.CommandLine;
 using System.CommandLine.NamingConventionBinder;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
-using Guppi.Core.Interfaces;
+using Guppi.Core.Interfaces.Services;
 using Spectre.Console;
 
 namespace Guppi.Console.Skills;
