+++ Guppi.Core/Interfaces/Services/IAsciiService.cs
@@ -1,6 +1,6 @@
 ﻿using Guppi.Core.Entities.Ascii;
 
-namespace Guppi.Core.Interfaces;
+namespace Guppi.Core.Interfaces.Services;
 
 public interface IAsciiService
 {
