+++ Guppi.Core/Interfaces/Services/IGitService.cs
@@ -1,4 +1,4 @@
-namespace Guppi.Core.Interfaces;
+namespace Guppi.Core.Interfaces.Services;
 
 public interface IGitService
 {
