+++ Guppi.Core/Interfaces/Services/ISerialPortService.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Guppi.Core.Interfaces;
+namespace Guppi.Core.Interfaces.Services;
 
 public interface ISerialPortService
 {
