+++ Guppi.Core/Interfaces/Services/ITodoService.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 
-namespace Guppi.Core.Interfaces;
+namespace Guppi.Core.Interfaces.Services;
 
 public interface ITodoService
 {
