+++ Guppi.Core/Providers/Calendar/ICalCalendarProvider.cs
@@ -5,9 +5,9 @@ using System.Threading.Tasks;
 using Guppi.Core;
 using Guppi.Core.Exceptions;
 using Guppi.Core.Entities.Calendar;
-using Guppi.Core.Interfaces;
 using Ical.Net.CalendarComponents;
 using Ical.Net.DataTypes;
+using Guppi.Core.Interfaces.Providers;
 
 namespace Guppi.Core.Providers.Calendar;
