+++ Guppi.Core/Providers/HttpRestProvider.cs
@@ -1,7 +1,7 @@
 using System.Net.Http;
 using System.Text.Json;
 using System.Threading.Tasks;
-using Guppi.Core.Interfaces;
+using Guppi.Core.Interfaces.Providers;
 
 #nullable enable
