+++ Guppi.Core/Services/WeatherService.cs
@@ -5,7 +5,8 @@ using Guppi.Core.Configurations;
 using Guppi.Core.Exceptions;
 using Guppi.Core.Services.Weather;
 using Guppi.Core.Entities.Weather;
-using Guppi.Core.Interfaces;
+using Guppi.Core.Interfaces.Providers;
+using Guppi.Core.Interfaces.Services;
 
 namespace Guppi.Core.Services;
